# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.Confirm:
    properties:
      id:
        type: string
      is payed:
        type: boolean
      status:
        type: string
    type: object
...
  dto.LatePaymentFeeReq:
    properties:
      name:
        type: string
      nominal:
        type: number
      unit:
        type: string
    type: object
...
  dto.LoanReq:
    properties:
      loan amount:
        type: integer
      loan handling cost id:
        type: string
      loan interest rate id:
        type: string
      loan late payment fess id:
        type: string
    type: object
  dto.Paging:
    properties:
      Page:
        type: integer
      size:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  dto.ResponseData:
    properties:
      data: {}
      message:
        type: string
    type: object
  dto.ResponseMessage:
    properties:
      message:
        type: string
    type: object
  dto.ResponsePaging:
    properties:
      data: {}
      message:
        type: string
      paging:
        $ref: '#/definitions/dto.Paging'
    type: object
...
  model.LatePaymentFee:
    properties:
      id:
        type: string
      name:
        type: string
      nominal:
        type: number
      unit:
        type: string
    type: object
...
info:
  contact: {}
  title: Laundry Apps
  version: "1.0"
paths:
...
  /latepaymentfee:
    post:
      consumes:
      - application/json
      description: create new data late fee payment
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data late fee
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LatePaymentFeeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseData'
      summary: new
      tags:
      - late fee payment
    put:
      consumes:
      - application/json
      description: update data late fee payment
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data late fee
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LatePaymentFee'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: update
      tags:
      - late fee payment
  /latepaymentfee/{id}:
    delete:
      consumes:
      - application/json
      description: delete data late fee payment
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: delete
      tags:
      - late fee payment
  /latepaymentfee/list/{page}/{size}:
    get:
      consumes:
      - application/json
      description: get all data late fee payment
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: page of pagination
        in: path
        name: page
        required: true
        type: integer
      - description: size of pagination
        in: path
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponsePaging'
      summary: get all
      tags:
      - late fee payment
  /loan:
    post:
      consumes:
      - application/json
      description: create new data loan
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data deposite interest
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoanReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: new
      tags:
      - loan
  /loan/{id}:
    get:
      consumes:
      - application/json
      description: find by id data loan
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: find by id loan
      tags:
      - loan
  /loan/confirm:
    put:
      consumes:
      - application/json
      description: confirm data uploaded
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Data late fee
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.Confirm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: confirm
      tags:
      - loan
  /loan/installment/:id:
    get:
      consumes:
      - application/json
      description: find by id data loan
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: find by id installenment
      tags:
      - installenment loan
  /loan/pay:
    post:
      consumes:
      - multipart/form-data
      description: upload recipe
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: File to upload
        in: formData
        name: file
        required: true
        type: file
      - description: ID parameter
        in: formData
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: upload
      tags:
      - loan
  /loan/updatedpayment:
    get:
      consumes:
      - application/json
      description: upload recipe
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: get data upoaded
      tags:
      - loan
  /loan/updatelatefee:
    get:
      consumes:
      - application/json
      description: create update data late fee
      parameters:
      - default: Bearer <token>
        description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ResponseMessage'
      summary: update
      tags:
      - loan
...
schemes:
- http
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"